+++ packages/expo-router/src/__tests__/prefetch.test.ios.tsx
@@ -961,7 +961,6 @@ it('can still use <Screen /> while prefetching in tabs', () => {
     'Should only change after focus',
     'index',
     'Should only change after focus',
-    'index',
   ]);
 });