+++ packages/expo-router/src/react-navigation/core/NavigationBuilderContext.tsx
@@ -8,18 +8,7 @@ export type ListenerMap = {
   focus: FocusedNavigationListener;
 };
 
-export type KeyedListenerMap = {
-  preventRemove: ChildPreventRemoveListener;
-  beforeRemove: ChildBeforeRemoveListener;
-};
-
 export type AddListener = <T extends keyof ListenerMap>(type: T, listener: ListenerMap[T]) => void;
-
-export type AddKeyedListener = <T extends keyof KeyedListenerMap>(
-  type: T,
-  key: string,
-  listener: KeyedListenerMap[T]
-) => void;
 
 export type FocusedNavigationCallback<T> = (navigation: NavigationHelpers<ParamListBase>) => T;
 
@@ -28,10 +17,6 @@ export type FocusedNavigationListener = <T>(callback: FocusedNavigationCallback<
   result: T;
 };
 
-export type ChildPreventRemoveListener = (action: NavigationAction) => boolean;
-
-export type ChildBeforeRemoveListener = (action: NavigationAction) => void;
-
 /**
  * Context which holds the required helpers needed to build nested navigators.
  */
@@ -39,12 +24,9 @@ export const NavigationBuilderContext = React.createContext<{
   handleAction: (action: NavigationAction, originKey?: string) => void;
   resetNavigator: (stateKey: string, routerType: string | undefined) => void;
   addListener?: AddListener;
-  addKeyedListener?: AddKeyedListener;
-  onDispatchAction: (action: NavigationAction, noop: boolean) => void;
   onOptionsChange: (options: object, routeKey?: string) => void;
 }>({
   handleAction: () => undefined,
   resetNavigator: () => undefined,
-  onDispatchAction: () => undefined,
   onOptionsChange: () => undefined,
 });